+++ components/appLayout.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Header from './header/header';
+import Header from './layout/header';
 
 const AppLayout = ({children}) => {
   return (
