+++ public/css/components/layout/header.module.css
@@ -44,7 +44,7 @@
   }
 }
 
-/* about */
+/* profile */
 
 @media (min-width: 1200px) {
   .header[data-type='about'] {
