--- public/css/style.css
@@ -1,9 +1,0 @@
-h1 {
-    color: red;
-}
-
-#background {
-    width: 500px;
-    height: 500px;
-    background-image: url("/images/zelda 001.jpg");
-}
